refactor(kernel::process): move from scheduler::process
