+++ crates/kernel/src/drivers/vfs/rodatafs/init_system.rs
@@ -1,7 +1,7 @@
 //! This module contains the boot-time implementiation of the initialization of the procfs.
 
 use super::{eve_journal::EVEJournal, usbinfo::USBInfoFile};
-use crate::scheduler::Pid;
+use crate::process::Pid;
 
 use super::{
     GenericRodFSFile, cpuinfo::CpuInfoFile, kernelinfo::KernelInfoFile, meminfo::MemInfoFile,
