+++ crates/kernel/src/scheduler/expose.rs
@@ -1,54 +1,12 @@
-use alloc::sync::Arc;
 use core::sync::atomic::Ordering;
 
 use crate::{
-    VirtAddr,
     arch::without_interrupts,
-    memory::paging::MapToError,
-    scheduler::{
-        SCHEDULER_INITED,
-        cpu_context::{Cid, ContextPriority, Thread},
-        this_process, this_thread,
-    },
-    time,
-    utils::types::Name,
-    warn,
-};
-use alloc::boxed::Box;
-use bitflags::bitflags;
-use safa_abi::{
-    self,
-    errors::ErrorStatus,
-    raw::{self, processes::ProcessStdio},
+    scheduler::{SCHEDULER_INITED, cpu_context::Cid, this_process, this_thread},
+    time, warn,
 };
 
-use crate::utils::path::{PathBuf, make_path};
-use thiserror::Error;
-
-use crate::{
-    drivers::vfs::{FSError, FSObjectType, FSResult, VFS_STRUCT, expose::File},
-    utils::{
-        elf::{Elf, ElfError},
-        io::Readable,
-        path::Path,
-    },
-};
-
-use super::{
-    Pid,
-    process::{Process, ProcessInfo},
-};
-
-#[unsafe(no_mangle)]
-pub fn process_exit(code: usize) -> ! {
-    without_interrupts(|| {
-        let current_process = super::this_process();
-        current_process.kill(code, None);
-
-        thread_yield();
-        unreachable!("process didn't exit")
-    })
-}
+use crate::process::{Pid, ProcessInfo};
 
 pub fn thread_exit(code: usize) -> ! {
     without_interrupts(|| {
@@ -191,218 +149,6 @@ pub fn getinfo(pid: Pid) -> Option<ProcessInfo> {
     super::find(|p| p.pid() == pid, |t| ProcessInfo::from(&**t))
 }
 
-bitflags! {
-    #[derive(Debug, Clone, Copy)]
-    pub struct SpawnFlags: u8 {
-        const CLONE_RESOURCES = 1 << 0;
-        const CLONE_CWD = 1 << 1;
-    }
-}
-
-impl From<raw::processes::SpawnFlags> for SpawnFlags {
-    fn from(value: raw::processes::SpawnFlags) -> Self {
-        unsafe { Self::from_bits_retain(core::mem::transmute(value)) }
-    }
-}
-
-#[derive(Debug, Clone, Error)]
-
-pub enum SpawnError {
-    #[error("out of memory")]
-    MapToError(#[from] MapToError),
-    #[error("failed to map elf to memory {0}")]
-    ElfError(#[from] ElfError),
-    #[error("error while creating process {0}")]
-    FSError(#[from] FSError),
-}
-
-#[inline(always)]
-fn spawn_inner(
-    name: Name,
-    flags: SpawnFlags,
-    stdio: ProcessStdio,
-    create_process: impl FnOnce(
-        Name,
-        Pid,
-        Pid,
-        Box<PathBuf>,
-    ) -> Result<(Arc<Process>, Arc<Thread>), SpawnError>,
-) -> Result<Pid, SpawnError> {
-    let this_process = this_process();
-    let this_state = this_process.state();
-
-    let cwd = if flags.contains(SpawnFlags::CLONE_CWD) {
-        this_state.cwd()
-    } else {
-        make_path!("ram", "")
-    };
-
-    let current_process = super::this_process();
-    let current_pid = current_process.pid();
-    let new_pid = super::SCHEDULER.write().add_pid();
-
-    let cwd = Box::new(cwd.into_owned().unwrap());
-    let (process, root_thread) = create_process(name, current_pid, new_pid, cwd)?;
-
-    let provide_resources = || {
-        let mut state = process.state_mut();
-        let Some(process_resources) = state.resource_manager_mut() else {
-            unreachable!();
-        };
-
-        drop(this_state);
-        let mut this_state = this_process.state_mut();
-
-        let clone = if flags.contains(SpawnFlags::CLONE_RESOURCES) {
-            this_state.clone_resources()
-        } else {
-            // clone only necessary resources
-            let mut resources = heapless::Vec::<usize, 3>::new();
-            if let Some(stdin) = stdio.stdin.into() {
-                _ = resources.push(stdin);
-            }
-
-            if let Some(stdout) = stdio.stdout.into() {
-                _ = resources.push(stdout);
-            }
-
-            if let Some(stderr) = stdio.stderr.into() {
-                _ = resources.push(stderr);
-            }
-
-            if resources.is_empty() {
-                return Ok(());
-            }
-            this_state.clone_specific_resources(&resources)?
-        };
-
-        process_resources.overwrite_resources(clone);
-        Ok(())
-    };
-
-    provide_resources().map_err(|()| FSError::InvalidResource)?;
-
-    let pid = super::add(process, root_thread);
-    Ok(pid)
-}
-
-fn spawn<T: Readable>(
-    name: Name,
-    reader: &T,
-    argv: &[&str],
-    env: &[&[u8]],
-    flags: SpawnFlags,
-    priority: ContextPriority,
-    stdio: ProcessStdio,
-    custom_stack_size: Option<usize>,
-) -> Result<Pid, SpawnError> {
-    spawn_inner(name, flags, stdio, |name: Name, ppid, pid, cwd| {
-        let elf = Elf::new(reader)?;
-        let process = Process::from_elf(
-            name,
-            pid,
-            ppid,
-            cwd,
-            elf,
-            argv,
-            env,
-            priority,
-            stdio,
-            custom_stack_size,
-        )?;
-        Ok(process)
-    })
-}
-
-/// spawns an elf process from a path
-pub fn pspawn(
-    name: Name,
-    path: Path,
-    argv: &[&str],
-    env: &[&[u8]],
-    flags: SpawnFlags,
-    priority: ContextPriority,
-    stdio: ProcessStdio,
-    custom_stack_size: Option<usize>,
-) -> Result<Pid, FSError> {
-    let file = File::open(path)?;
-
-    if file.kind() != FSObjectType::File {
-        return Err(FSError::NotAFile);
-    }
-
-    spawn(
-        name,
-        &file,
-        argv,
-        env,
-        flags,
-        priority,
-        stdio,
-        custom_stack_size,
-    )
-    .map_err(|_| FSError::NotExecutable)
-}
-
-/// Spawns a userspace function in a new thread
-///
-/// if `cpu` is Some it will append to that CPU instead of choosing one, use CPU 0 to append to boot CPU
-pub fn thread_spawn(
-    entry_point: VirtAddr,
-    argument_ptr: VirtAddr,
-    priority: Option<ContextPriority>,
-    cpu: Option<usize>,
-    custom_stack_size: Option<usize>,
-) -> Result<Cid, MapToError> {
-    let this = this_process();
-    let (thread, cid) = Process::add_thread_to_process(
-        &this,
-        entry_point,
-        argument_ptr,
-        priority,
-        custom_stack_size,
-    )?;
-    super::add_thread(thread, cpu);
-    Ok(cid)
-}
-
-/// Spawns a kernel function in a new thread
-///
-/// if `cpu` is Some it will append to that CPU instead of choosing one, use CPU 0 to append to boot CPU
-pub fn kernel_thread_spawn<T: 'static>(
-    func: fn(cid: Cid, &'static T) -> !,
-    arg: &'static T,
-    priority: Option<ContextPriority>,
-    cpu: Option<usize>,
-) -> Result<Cid, MapToError> {
-    thread_spawn(
-        VirtAddr::from(func as usize),
-        VirtAddr::from(arg as *const T as usize),
-        priority,
-        cpu,
-        None,
-    )
-}
-
-/// also ensures the cwd ends with /
-/// will only Err if new_dir doesn't exists or is not a directory
-#[unsafe(no_mangle)]
-pub fn chdir(new_dir: Path) -> FSResult<()> {
-    VFS_STRUCT.read().verify_path_dir(new_dir)?;
-
-    let process = this_process();
-    let mut state = process.state_mut();
-    let cwd = state.cwd_mut();
-
-    if new_dir.is_absolute() {
-        *cwd = new_dir.into_owned_simple()?;
-    } else {
-        cwd.append_simplified(new_dir)?;
-    }
-
-    Ok(())
-}
-
 fn can_terminate(mut process_ppid: Pid, process_pid: Pid, terminator_pid: Pid) -> bool {
     if process_ppid == terminator_pid || process_pid == terminator_pid {
         return true;
@@ -458,14 +204,4 @@ pub fn pkill(pid: Pid) -> Result<(), ()> {
         return Ok(());
     }
     Err(())
-}
-
-#[unsafe(no_mangle)]
-/// extends program break by `amount`
-/// returns the new program break ptr
-/// on fail returns null
-pub fn sbrk(amount: isize) -> Result<*mut u8, ErrorStatus> {
-    let current_process = super::this_process();
-    let mut state = current_process.state_mut();
-    state.extend_data_by(amount).ok_or(ErrorStatus::OutOfMemory)
 }
