+++ crates/kernel/src/scheduler/queue.rs
@@ -2,7 +2,7 @@ use core::{marker::PhantomData, ptr::NonNull};
 
 use alloc::{boxed::Box, sync::Arc};
 
-use crate::scheduler::process::Process;
+use crate::process::Process;
 
 struct Node<T> {
     inner: T,
