+++ crates/kernel/src/scheduler/tests.rs
@@ -1,20 +1,18 @@
-use crate::utils::path::make_path;
+use crate::process::spawn::SpawnFlags;
 use crate::utils::types::Name;
+use crate::{process, utils::path::make_path};
 
 use safa_abi::raw::processes::ProcessStdio;
 
 use crate::{
     arch::without_interrupts,
-    scheduler::{
-        cpu_context::ContextPriority,
-        expose::{SpawnFlags, pspawn, wait_for_process},
-    },
+    scheduler::{cpu_context::ContextPriority, expose::wait_for_process},
 };
 
 #[test_case]
 fn spawn_test() {
     without_interrupts(|| {
-        let pid = pspawn(
+        let pid = process::spawn::pspawn(
             Name::try_from("TEST_CASE").unwrap(),
             make_path!("sys", "/bin/true"),
             &[],
