+++ crates/kernel/src/syscalls/mod.rs
@@ -1,8 +1,10 @@
 use safa_abi::errors::{ErrorStatus, SysResult};
 
 use crate::drivers::vfs::expose::FileAttr;
+use crate::process::Pid;
 use crate::scheduler::cpu_context::Cid;
-use crate::scheduler::{self, Pid, resources};
+use crate::scheduler::{self, resources};
+
 use crate::time;
 use crate::utils::syscalls::{SyscallFFI, SyscallTable};
 use crate::{
@@ -85,7 +87,7 @@ pub fn syscall(number: u16, a: usize, b: usize, c: usize, d: usize, e: usize) ->
                 processes::syspspawn_raw((a as *const u8, b), c as *const _, d as *mut Pid)
             }
             SyscallTable::SysTSpawn => processes::sys_tspawn_raw(a, b as *const _, c as *mut Cid),
-            SyscallTable::SysPExit => scheduler::expose::process_exit(a),
+            SyscallTable::SysPExit => crate::process::current::exit(a),
             SyscallTable::SysTExit => scheduler::expose::thread_exit(a),
             SyscallTable::SysTYield => Ok(scheduler::expose::thread_yield()),
             SyscallTable::SysTSleep => {
